+++ packages/client/src/pages/Game/Game.module.scss
@@ -1,11 +1,7 @@
 .wrapper {
-  /* background-color: white; */
   padding: 0 24px;
-  /* border-radius: 24px; */
   display: grid;
   grid-template-columns: 3fr 1fr;
-  /* width: calc(100% - 128px); */
-  /* height: calc(100% - 96px); */
   gap: 24px;
 }
